Adds strong names for all assemblies using a new strong name key
In order to reference NConsole.dll by other assemblies with strong
names, the assemblies in this project also need strong names.  For the
convenience of consumers of this project, a new strong name key is
included named NConsole.snk.  This key is brand new and has not be used
for any other purpose.

The existence of this strong name key within the repository itself
implies, of course, that the key is not secure.  If a secure version of
the assembly is needed, you must sign these assemblies with your own
key.